Add hemtt as build tool
