--- addons/grad_slinghelmet/Stringtable.xml
@@ -1,19 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-<Project name="grad_slingHelmet">
-	<Package name="Mission One">
-		<Container name="Misc">
-			<Key ID="str_grad_slingHelmet_sling">
-				<Original>Sling Helmet</Original>
-				<English>Sling Helmet</English>
-				<German>Helm absetzen</German>
-				<Japanese>ヘルメットをかける</Japanese>
-			</Key>
-               <Key ID="str_grad_slingHelmet_unsling">
-				<Original>Unsling Helmet</Original>
-				<English>Unsling Helmet</English>
-				<German>Helm aufsetzen</German>
-				<Japanese>ヘルメットを被る</Japanese>
-			</Key>
-		</Container>
-	</Package>
-</Project>
