--- addons/grad_slinghelmet/allowedHeadgear.hpp
@@ -1,758 +1,0 @@
-private _allowedHeadgear = [
-
-  // Vanilla + Apex + Laws Of War
-  "H_HelmetSpecO_blk",
-  "H_HelmetSpecO_ocamo",
-  "H_HelmetB",
-  "H_HelmetB_black",
-  "H_HelmetB_camo",
-  "H_HelmetB_desert",
-  "H_HelmetB_grass",
-  "H_HelmetB_sand",
-  "H_HelmetB_snakeskin",
-  "H_HelmetCrew_I",
-  "H_HelmetCrew_O",
-  "H_HelmetCrew_B",
-  "H_HelmetLeaderO_ocamo",
-  "H_HelmetLeaderO_oucamo",
-  "H_HelmetSpecB",
-  "H_HelmetSpecB_blk",
-  "H_HelmetSpecB_paint2",
-  "H_HelmetSpecB_paint1",
-  "H_HelmetSpecB_sand",
-  "H_HelmetSpecB_snakeskin",
-  "H_CrewHelmetHeli_I",
-  "H_CrewHelmetHeli_O",
-  "H_CrewHelmetHeli_B",
-  "H_PilotHelmetHeli_I",
-  "H_PilotHelmetHeli_O",
-  "H_PilotHelmetHeli_B",
-  "H_HelmetB_light",
-  "H_HelmetB_light_black",
-  "H_HelmetB_light_desert",
-  "H_HelmetB_light_grass",
-  "H_HelmetB_light_sand",
-  "H_HelmetB_light_snakeskin",
-  "H_HelmetIA",
-  "H_HelmetO_ocamo",
-  "H_HelmetO_oucamo",
-  "H_RacingHelmet_1_black_F",
-  "H_RacingHelmet_1_blue_F",
-  "H_RacingHelmet_2_F",
-  "H_RacingHelmet_1_F",
-  "H_RacingHelmet_1_green_F",
-  "H_RacingHelmet_1_orange_F",
-  "H_RacingHelmet_1_red_F",
-  "H_RacingHelmet_3_F",
-  "H_RacingHelmet_4_F",
-  "H_RacingHelmet_1_white_F",
-  "H_RacingHelmet_1_yellow_F",
-  "H_HelmetSpecO_ghex_F",
-  "H_HelmetB_tna_F",
-  "H_HelmetCrew_O_ghex_F",
-  "H_HelmetLeaderO_ghex_F",
-  "H_HelmetB_Enh_tna_F",
-  "H_HelmetB_Light_tna_F",
-  "H_HelmetO_ghex_F",
-  "H_Helmet_Skate",
-  "H_PASGT_neckprot_blue_press_F",
-  "H_PASGT_neckprot_blue_press_F",
-  "H_Construction_headset_yellow_F",
-  "H_Construction_earprot_yellow_F",
-  "H_Construction_basic_yellow_F",
-  "H_Construction_headset_white_F",
-  "H_Construction_earprot_white_F",
-  "H_Construction_basic_white_F",
-  "H_Construction_headset_vrana_F",
-  "H_Construction_earprot_vrana_F",
-  "H_Construction_basic_vrana_F",
-  "H_Construction_headset_red_F",
-  "H_Construction_earprot_red_F",
-  "H_Construction_basic_red_F",
-  "H_Construction_headset_orange_F",
-  "H_Construction_earprot_orange_F",
-  "H_Construction_basic_orange_F",
-  "H_Construction_headset_black_F",
-  "H_Construction_earprot_black_F",
-  "H_Construction_basic_black_F",
-  "H_PASGT_basic_white_F",
-  "H_PASGT_basic_olive_F",
-  "H_PASGT_basic_blue_F",
-  "H_PASGT_basic_black_F",
-
-  // SMA
-  "SMA_Helmet_Tan",
-  "SMA_Helmet_Green",
-  "SMA_Helmet_Black",
-
-  // ARC GER
-  "ARC_GER_Tropentarn_Mich",
-  "ARC_GER_Flecktarn_Mich",
-  "ARC_GER_Tropentarn_Helmet_simple_medic",
-  "ARC_GER_Flecktarn_Helmet_simple_Medic",
-  "ARC_GER_Tropentarn_Helmet_simple",
-  "ARC_GER_Flecktarn_Helmet_simple",
-  "ARC_GER_Tropentarn_Helmet_Light_Medic",
-  "ARC_GER_Flecktarn_Helmet_Medic",
-  "ARC_GER_Tropentarn_Helmet_Light_EDO",
-  "ARC_GER_Flecktarn_Helmet_EOD",
-  "ARC_GER_Tropentarn_Helmet_Light",
-  "ARC_GER_Flecktarn_Helmet_Light",
-  "ARC_GER_Tropentarn_Helmet",
-  "ARC_GER_Flecktarn_Helmet",
-
-  // PROJECT OPFOR
-  "LOP_H_SSh68Helmet_TAN",
-  "LOP_H_SSh68Helmet_OLV",
-  "LOP_H_SSh68Helmet_BLK",
-  "PO_H_mich_bare_isof_3_nortos_hs_blk",
-  "PO_H_mich_bare_isof_3_nortos_alt_blk",
-  "PO_H_mich_bare_isof_3_nortos_blk",
-  "PO_H_mich_bare_isof_2_nortos_hs_blk",
-  "PO_H_mich_bare_isof_2_nortos_alt_blk",
-  "PO_H_mich_bare_isof_2_nortos_blk",
-  "PO_H_mich_bare_isof_1_nortos_hs_blk",
-  "PO_H_mich_bare_isof_1_nortos_alt_blk",
-  "PO_H_mich_bare_isof_1_nortos_blk",
-  "LOP_H_6B27M_ess_wz93",
-  "LOP_H_6B27M_wz93",
-  "LOP_H_6B27M_ess_UN",
-  "LOP_H_6B27M_UN",
-  "LOP_H_6B27M_ess_UKR",
-  "LOP_H_6B27M_UKR",
-  "LOP_H_6B27M_ess_TRI",
-  "LOP_H_6B27M_TRI",
-  "LOP_H_6B27M_ess_Skol",
-  "LOP_H_6B27M_Skol",
-  "LOP_H_6B27M_ess_RACS",
-  "LOP_H_6B27M_RACS",
-  "LOP_H_6B27M_ess_MTP",
-  "LOP_H_6B27M_MTP",
-  "LOP_H_6B27M_ess_WDL",
-  "LOP_H_6B27M_WDL",
-  "LOP_H_6B27M_ess_M81",
-  "LOP_H_6B27M_M81",
-  "LOP_H_6B27M_ess_Izlom",
-  "LOP_H_6B27M_Izlom",
-  "LOP_H_6B27M_ess_IDPM",
-  "LOP_H_6B27M_IDPM",
-  "LOP_H_6B27M_ess_Flora",
-  "LOP_H_6B27M_Flora",
-  "LOP_H_6B27M_ess_Digit",
-  "LOP_H_6B27M_Digit",
-  "LOP_H_6B27M_ess_CDF",
-  "LOP_H_6B27M_CDF",
-  "LOP_H_6B27M_ess_DBLU",
-  "LOP_H_6B27M_DBLU",
-  "LOP_H_6B27M_ess_BLK",
-  "LOP_H_6B27M_BLK",
-  "LOP_H_6B27M_ess_ANA",
-  "LOP_H_6B27M_ANA",
-
-  // RHS: SAF
-  "LOP_H_6B27M_ANA",
-  "rhssaf_helmet_m97_woodland_black_ess",
-  "rhssaf_helmet_m97_woodland",
-  "rhssaf_helmet_m97_olive_nocamo_black_ess_bare",
-  "rhssaf_helmet_m97_olive_nocamo_black_ess",
-  "rhssaf_helmet_m97_olive_nocamo",
-  "rhssaf_helmet_m97_oakleaf_black_ess_bare",
-  "rhssaf_helmet_m97_oakleaf_black_ess",
-  "rhssaf_helmet_m97_oakleaf",
-  "rhssaf_helmet_m97_md2camo_black_ess_bare",
-  "rhssaf_helmet_m97_md2camo_black_ess",
-  "rhssaf_helmet_m97_md2camo",
-  "rhssaf_helmet_m97_digital_black_ess_bare",
-  "rhssaf_helmet_m97_digital_black_ess",
-  "rhssaf_helmet_m97_digital",
-  "rhssaf_helmet_m97_nostrap_blue_tan_ess_bare",
-  "rhssaf_helmet_m97_nostrap_blue_tan_ess",
-  "rhssaf_helmet_m97_nostrap_blue",
-  "rhssaf_helmet_m97_black_nocamo_black_ess_bare",
-  "rhssaf_helmet_m97_black_nocamo_black_ess",
-  "rhssaf_helmet_m97_black_nocamo",
-  "rhssaf_helmet_m59_85_oakleaf",
-  "rhssaf_helmet_m59_85_nocamo",
-
-  // RHS: GREF
-  "rhsgref_6b27m_ttsko_digi",
-  "rhsgref_6b27m_ttsko_forest",
-  "rhsgref_6b27m_ttsko_mountain",
-  "rhsgref_6b27m_ttsko_urban",
-  "rhsgref_helmet_M1_bare",
-  "rhsgref_helmet_M1_bare_alt01",
-  "rhsgref_helmet_M1_painted",
-  "rhsgref_helmet_M1_painted_alt01",
-  "rhsgref_M56",
-  "rhsgref_helmet_pasgt_3color_desert",
-  "rhsgref_helmet_pasgt_3color_desert_rhino",
-  "rhsgref_helmet_pasgt_altis_lizard",
-  "rhsgref_helmet_pasgt_erdl",
-  "rhsgref_helmet_pasgt_erdl_rhino",
-  "rhsgref_helmet_pasgt_flecktarn",
-  "rhsgref_helmet_pasgt_olive",
-  "rhsgref_helmet_pasgt_press",
-  "rhsgref_helmet_pasgt_un",
-  "rhsgref_helmet_pasgt_woodland",
-  "rhsgref_helmet_pasgt_woodland_rhino",
-  "rhsgref_ssh68_emr",
-  "rhsgref_ssh68_ttsko_digi",
-  "rhsgref_ssh68_ttsko_forest",
-  "rhsgref_ssh68_ttsko_mountain",
-  "rhsgref_ssh68_ttsko_dark",
-  "rhsgref_ssh68_un",
-  "rhsgref_ssh68_vsr",
-
-  // RHS: AFRF
-  "rhs_6b26_green",
-  "rhs_6b26_green",
-  "rhs_6b26_ess_green",
-  "rhs_6b26_ess_bala_green",
-  "rhs_6b26",
-  "rhs_6b26_bala",
-  "rhs_6b26_ess",
-  "rhs_6b26_ess_bala",
-  "rhs_6b27m_green",
-  "rhs_6b27m_green_ess",
-  "rhs_6b27m_digi",
-  "rhs_6b27m_digi_ess",
-  "rhs_6b27m",
-  "rhs_6b27m_ess",
-  "rhs_6b27m_ml",
-  "rhs_6b27m_ml_ess",
-  "rhs_6b28_green",
-  "rhs_6b28_green_ess",
-  "rhs_6b28",
-  "rhs_6b28_ess",
-  "rhs_6b28_flora",
-  "rhs_6b28_flora_ess",
-  "rhs_6b47",
-  "rhs_6b47_ess",
-  "rhs_6b7_1m",
-  "rhs_6b7_1m_emr",
-  "rhs_6b7_1m_emr_ess",
-  "rhs_6b7_1m_ess",
-  "rhs_6b7_1m_flora",
-  "rhs_6b7_1m_flora_ns3",
-  "rhs_6b7_1m_olive",
-  "rhs_altyn_novisor",
-  "rhs_altyn_novisor_ess",
-  "rhs_altyn_visordown",
-  "rhs_altyn",
-  "rhs_ssh68",
-  "rhs_zsh7a_mike_alt",
-  "rhs_zsh7a_mike_green_alt",
-  "rhs_zsh7a_mike_green",
-  "rhs_zsh7a_mike",
-
-  //RHS: USAF
-  "rhsusf_ach_bare",
-  "rhsusf_ach_bare_des",
-  "rhsusf_ach_bare_des_ess",
-  "rhsusf_ach_bare_des_headset",
-  "rhsusf_ach_bare_des_headset_ess",
-  "rhsusf_ach_bare_ess",
-  "rhsusf_ach_bare_headset",
-  "rhsusf_ach_bare_headset_ess",
-  "rhsusf_ach_bare_semi",
-  "rhsusf_ach_bare_semi_ess",
-  "rhsusf_ach_bare_semi_headset",
-  "rhsusf_ach_bare_semi_headset_ess",
-  "rhsusf_ach_bare_tan",
-  "rhsusf_ach_bare_tan_ess",
-  "rhsusf_ach_bare_tan_headset",
-  "rhsusf_ach_bare_tan_headset_ess",
-  "rhsusf_ach_bare_wood",
-  "rhsusf_ach_bare_wood_ess",
-  "rhsusf_ach_bare_wood_headset",
-  "rhsusf_ach_bare_wood_headset_ess",
-  "rhsusf_ach_helmet_M81",
-  "rhsusf_ach_helmet_ocp",
-  "rhsusf_ach_helmet_ESS_ocp",
-  "rhsusf_ach_helmet_headset_ocp",
-  "rhsusf_ach_helmet_headset_ess_ocp",
-  "rhsusf_ach_helmet_camo_ocp",
-  "rhsusf_ach_helmet_ocp_norotos",
-  "rhsusf_ach_helmet_ucp",
-  "rhsusf_ach_helmet_ESS_ucp",
-  "rhsusf_ach_helmet_headset_ucp",
-  "rhsusf_ach_helmet_headset_ess_ucp",
-  "rhsusf_ach_helmet_ucp_norotos",
-  "rhsusf_cvc_green_helmet",
-  "rhsusf_cvc_green_alt_helmet",
-  "rhsusf_cvc_green_ess",
-  "rhsusf_cvc_helmet",
-  "rhsusf_cvc_alt_helmet",
-  "rhsusf_cvc_ess",
-  "rhsusf_opscore_aor1",
-  "rhsusf_opscore_aor1_pelt",
-  "rhsusf_opscore_aor1_pelt_nsw",
-  "rhsusf_opscore_aor2",
-  "rhsusf_opscore_aor2_pelt",
-  "rhsusf_opscore_aor2_pelt_nsw",
-  "rhsusf_opscore_bk",
-  "rhsusf_opscore_bk_pelt",
-  "rhsusf_opscore_coy_cover",
-  "rhsusf_opscore_coy_cover_pelt",
-  "rhsusf_opscore_fg",
-  "rhsusf_opscore_fg_pelt",
-  "rhsusf_opscore_fg_pelt_cam",
-  "rhsusf_opscore_fg_pelt_nsw",
-  "rhsusf_opscore_mc_cover",
-  "rhsusf_opscore_mc_cover_pelt",
-  "rhsusf_opscore_mc_cover_pelt_nsw",
-  "rhsusf_opscore_mc_cover_pelt_cam",
-  "rhsusf_opscore_mc",
-  "rhsusf_opscore_mc_pelt",
-  "rhsusf_opscore_mc_pelt_nsw",
-  "rhsusf_opscore_paint",
-  "rhsusf_opscore_paint_pelt",
-  "rhsusf_opscore_paint_pelt_nsw",
-  "rhsusf_opscore_paint_pelt_nsw_cam",
-  "rhsusf_opscore_rg_cover",
-  "rhsusf_opscore_rg_cover_pelt",
-  "rhsusf_opscore_ut",
-  "rhsusf_opscore_ut_pelt",
-  "rhsusf_opscore_ut_pelt_cam",
-  "rhsusf_opscore_ut_pelt_nsw",
-  "rhsusf_opscore_ut_pelt_nsw_cam",
-  "rhsusf_opscore_mar_fg",
-  "rhsusf_opscore_mar_fg_pelt",
-  "rhsusf_opscore_mar_ut",
-  "rhsusf_opscore_mar_ut_pelt",
-  "rhsusf_hgu56p_black",
-  "rhsusf_hgu56p_visor_black",
-  "rhsusf_hgu56p_green",
-  "rhsusf_hgu56p_visor_green",
-  "rhsusf_hgu56p",
-  "rhsusf_hgu56p_visor",
-  "rhsusf_hgu56p_saf",
-  "rhsusf_hgu56p_visor_saf",
-  "rhsusf_hgu56p_usa",
-  "rhsusf_hgu56p_visor_usa",
-  "rhsusf_lwh_helmet_M1942",
-  "rhsusf_lwh_helmet_marpatd",
-  "rhsusf_lwh_helmet_marpatd_ess",
-  "rhsusf_lwh_helmet_marpatd_headset",
-  "rhsusf_lwh_helmet_marpatwd",
-  "rhsusf_lwh_helmet_marpatwd_blk_ess",
-  "rhsusf_lwh_helmet_marpatwd_headset_blk",
-  "rhsusf_lwh_helmet_marpatwd_headset",
-  "rhsusf_lwh_helmet_marpatwd_ess",
-  "rhsusf_mich_bare",
-  "rhsusf_mich_bare_alt",
-  "rhsusf_mich_bare_headset",
-  "rhsusf_mich_bare_norotos",
-  "rhsusf_mich_bare_norotos_alt",
-  "rhsusf_mich_bare_norotos_alt_headset",
-  "rhsusf_mich_bare_norotos_arc",
-  "rhsusf_mich_bare_norotos_arc_alt",
-  "rhsusf_mich_bare_norotos_arc_alt_headset",
-  "rhsusf_mich_bare_norotos_arc_headset",
-  "rhsusf_mich_bare_norotos_headset",
-  "rhsusf_mich_bare_semi",
-  "rhsusf_mich_bare_alt_semi",
-  "rhsusf_mich_bare_semi_headset",
-  "rhsusf_mich_bare_norotos_semi",
-  "rhsusf_mich_bare_norotos_alt_semi",
-  "rhsusf_mich_bare_norotos_alt_semi_headset",
-  "rhsusf_mich_bare_norotos_arc_semi",
-  "rhsusf_mich_bare_norotos_arc_alt_semi",
-  "rhsusf_mich_bare_norotos_arc_alt_semi_headset",
-  "rhsusf_mich_bare_norotos_arc_semi_headset",
-  "rhsusf_mich_bare_norotos_semi_headset",
-  "rhsusf_mich_bare_tan",
-  "rhsusf_mich_bare_alt_tan",
-  "rhsusf_mich_bare_tan_headset",
-  "rhsusf_mich_bare_norotos_tan",
-  "rhsusf_mich_bare_norotos_alt_tan",
-  "rhsusf_mich_bare_norotos_alt_tan_headset",
-  "rhsusf_mich_bare_norotos_arc_tan",
-  "rhsusf_mich_bare_norotos_arc_alt_tan",
-  "rhsusf_mich_bare_norotos_arc_alt_tan_headset",
-  "rhsusf_mich_bare_norotos_tan_headset",
-  "rhsusf_mich_helmet_marpatd",
-  "rhsusf_mich_helmet_marpatd_alt",
-  "rhsusf_mich_helmet_marpatd_alt_headset",
-  "rhsusf_mich_helmet_marpatd_headset",
-  "rhsusf_mich_helmet_marpatd_norotos",
-  "rhsusf_mich_helmet_marpatd_norotos_arc",
-  "rhsusf_mich_helmet_marpatd_norotos_arc_headset",
-  "rhsusf_mich_helmet_marpatd_norotos_headset",
-  "rhsusf_mich_helmet_marpatwd",
-  "rhsusf_mich_helmet_marpatwd_alt",
-  "rhsusf_mich_helmet_marpatwd_alt_headset",
-  "rhsusf_mich_helmet_marpatwd_headset",
-  "rhsusf_mich_helmet_marpatwd_norotos",
-  "rhsusf_mich_helmet_marpatwd_norotos_arc",
-  "rhsusf_mich_helmet_marpatwd_norotos_arc_headset",
-  "rhsusf_mich_helmet_marpatwd_norotos_headset",
-  "rhsusf_protech_helmet",
-  "rhsusf_protech_helmet_ess",
-  "rhsusf_protech_helmet_rhino",
-  "rhsusf_protech_helmet_rhino_ess",
-
-  //3CB
-  "UK3CB_BAF_H_CrewHelmet_DDPM_A",
-  "UK3CB_BAF_H_CrewHelmet_DDPM_ESS_A",
-  "UK3CB_BAF_H_CrewHelmet_DPMT_ESS_A",
-  "UK3CB_BAF_H_CrewHelmet_DPMT_A",
-  "UK3CB_BAF_H_CrewHelmet_DPMW_A",
-  "UK3CB_BAF_H_CrewHelmet_DPMW_ESS_A",
-  "UK3CB_BAF_H_CrewHelmet_A",
-  "UK3CB_BAF_H_CrewHelmet_ESS_A",
-  "UK3CB_BAF_H_CrewHelmet_B",
-  "UK3CB_BAF_H_PilotHelmetHeli_A",
-  "UK3CB_BAF_H_Mk6_DDPM_A",
-  "UK3CB_BAF_H_Mk6_DDPM_B",
-  "UK3CB_BAF_H_Mk6_DDPM_C",
-  "UK3CB_BAF_H_Mk6_DDPM_D",
-  "UK3CB_BAF_H_Mk6_DDPM_E",
-  "UK3CB_BAF_H_Mk6_DDPM_F",
-  "UK3CB_BAF_H_Mk6_DPMT_A",
-  "UK3CB_BAF_H_Mk6_DPMT_B",
-  "UK3CB_BAF_H_Mk6_DPMT_C",
-  "UK3CB_BAF_H_Mk6_DPMT_D",
-  "UK3CB_BAF_H_Mk6_DPMT_E",
-  "UK3CB_BAF_H_Mk6_DPMT_F",
-  "UK3CB_BAF_H_Mk6_DPMW_A",
-  "UK3CB_BAF_H_Mk6_DPMW_B",
-  "UK3CB_BAF_H_Mk6_DPMW_C",
-  "UK3CB_BAF_H_Mk6_DPMW_D",
-  "UK3CB_BAF_H_Mk6_DPMW_E",
-  "UK3CB_BAF_H_Mk6_DPMW_F",
-  "UK3CB_BAF_H_Mk7_Win_A",
-  "UK3CB_BAF_H_Mk7_Win_ESS_A",
-  "UK3CB_BAF_H_Mk7_HiVis",
-  "UK3CB_BAF_H_Mk7_Camo_A",
-  "UK3CB_BAF_H_Mk7_Camo_B",
-  "UK3CB_BAF_H_Mk7_Camo_C",
-  "UK3CB_BAF_H_Mk7_Camo_E",
-  "UK3CB_BAF_H_Mk7_Camo_F",
-  "UK3CB_BAF_H_Mk7_Camo_CESS_A",
-  "UK3CB_BAF_H_Mk7_Camo_CESS_B",
-  "UK3CB_BAF_H_Mk7_Camo_CESS_C",
-  "UK3CB_BAF_H_Mk7_Camo_ESS_A",
-  "UK3CB_BAF_H_Mk7_Camo_ESS_B",
-  "UK3CB_BAF_H_Mk7_Camo_ESS_C",
-  "UK3CB_BAF_H_Mk7_Net_A",
-  "UK3CB_BAF_H_Mk7_Net_B",
-  "UK3CB_BAF_H_Mk7_Net_C",
-  "UK3CB_BAF_H_Mk7_Net_CESS_A",
-  "UK3CB_BAF_H_Mk7_Net_CESS_B",
-  "UK3CB_BAF_H_Mk7_Net_CESS_C",
-  "UK3CB_BAF_H_Mk7_Net_ESS_A",
-  "UK3CB_BAF_H_Mk7_Net_ESS_B",
-  "UK3CB_BAF_H_Mk7_Net_ESS_C",
-  "UK3CB_BAF_H_Mk7_Scrim_A",
-  "UK3CB_BAF_H_Mk7_Scrim_B",
-  "UK3CB_BAF_H_Mk7_Scrim_C",
-  "UK3CB_BAF_H_Mk7_Scrim_D",
-  "UK3CB_BAF_H_Mk7_Scrim_E",
-  "UK3CB_BAF_H_Mk7_Scrim_F",
-  "UK3CB_BAF_H_Mk7_Scrim_ESS_A",
-  "UK3CB_BAF_H_Mk7_Scrim_ESS_B",
-  "UK3CB_BAF_H_Mk7_Scrim_ESS_C",
-
-  //BW Mod
-  "BWA3_CrewmanKSK_Schwarz_Headset",
-  "BWA3_CrewmanKSK_Schwarz",
-  "BWA3_CrewmanKSK_Fleck_Headset",
-  "BWA3_CrewmanKSK_Fleck",
-  "BWA3_CrewmanKSK_Tropen_Headset",
-  "BWA3_CrewmanKSK_Tropen",
-  "BWA3_Knighthelm",
-  "BWA3_M92_Fleck",
-  "BWA3_M92_Tropen",
-  "BWA3_MICH_Fleck",
-  "BWA3_MICH_Tropen",
-  "BWA3_OpsCore_Schwarz_Camera",
-  "BWA3_OpsCore_Schwarz",
-  "BWA3_OpsCore_Fleck_Camera",
-  "BWA3_OpsCore_Fleck_Patch",
-  "BWA3_OpsCore_Fleck",
-  "BWA3_OpsCore_Tropen_Camera",
-  "BWA3_OpsCore_Tropen_Patch",
-  "BWA3_OpsCore_Tropen",
-
-  //R3F
-  "R3F_casque_spectra_EOD",
-  "R3F_casque_spectra_EOD_DA",
-  "R3F_casque_equipage",
-  "R3F_casqueFS_olive",
-  "R3F_casqueFS_tan",
-  "R3F_casqueFS_noir",
-  "R3F_casqueFS_SNOW",
-  "R3F_casque_pilote",
-  "R3F_casque_spectra",
-  "R3F_casque_spectra_bleu",
-  "R3F_casque_spectra_jaune",
-  "R3F_casque_spectra_violet",
-  "R3F_casque_spectra_orange",
-  "R3F_casque_spectra_vert",
-  "R3F_casque_spectra_DA",
-  "R3F_casque_spectra_DA_bleu",
-  "R3F_casque_spectra_DA_jaune",
-  "R3F_casque_spectra_DA_violet",
-  "R3F_casque_spectra_DA_orange",
-  "R3F_casque_spectra_DA_vert",
-  "R3F_casque_spectra_SNOW",
-
-  //FoW WW2
-  "fow_h_ger_m40_heer_camo",
-  "fow_h_ger_m40_heer_01",
-  "fow_h_ger_m40_heer_02",
-  "fow_h_ger_m40_heer_01_net",
-  "fow_h_ger_m42_heer_01",
-  "fow_h_ger_m40_luftwaffe_01",
-  "fow_h_ger_m40_luftwaffe_02",
-  "fow_h_ger_m40_fall_01",
-  "fow_h_ger_m40_fall_01_net",
-  "fow_h_ger_m40_fall_01_camo",
-  "fow_h_ger_m40_fall_02_camo",
-  "fow_h_ger_m40_ss_01",
-  "fow_h_ger_m40_ss_02",
-  "fow_h_ger_feldmutze",
-  "fow_h_ger_feldmutze_ss",
-  "fow_h_ger_feldmutze_panzer",
-  "fow_h_ger_m38_feldmutze",
-  "fow_h_ger_m38_feldmutze_panzer",
-  "fow_h_ger_officer_cap",
-  "fow_h_ger_officer_cap_ss",
-  "fow_h_ger_headset",
-  "fow_h_ija_type90",
-  "fow_h_ija_type90_net",
-  "fow_h_ija_type90_net_neck",
-  "fow_h_ija_type90_foliage",
-  "fow_h_ija_type90_snlf",
-  "fow_h_ija_tank_helmet",
-  "fow_h_ija_flight_helmet",
-  "fow_h_ija_fieldcap",
-  "fow_h_ija_fieldcap_neck",
-  "fow_h_ija_fieldcap_officer",
-  "fow_h_ija_fieldcap_marine",
-  "fow_h_ija_fieldcap_marine_neck",
-  "fow_h_ija_hakimachi",
-  "fow_h_hi_beret",
-  "fow_h_hi_beret_headset",
-  "fow_h_hi_officer_cap",
-  "fow_h_us_m1",
-  "fow_h_us_m1_closed",
-  "fow_h_us_m1_folded",
-  "fow_h_us_m1_net",
-  "fow_h_us_m1_medic",
-  "fow_h_us_m1_officer",
-  "fow_h_us_m1_officer_closed",
-  "fow_h_us_m1_officer_folded",
-  "fow_h_us_m1_nco",
-  "fow_h_us_m1_nco_closed",
-  "fow_h_us_m1_nco_folded",
-  "fow_h_us_m2",
-  "fow_h_us_m2_ace",
-  "fow_h_us_m2_heart",
-  "fow_h_us_m2_diamond",
-  "fow_h_us_m2_clover",
-  "fow_h_us_m2_net",
-  "fow_h_us_m2_net_ace",
-  "fow_h_us_m2_net_heart",
-  "fow_h_us_m2_net_diamond",
-  "fow_h_us_m2_net_clover",
-  "fow_h_us_m2_camo",
-  "fow_h_us_m2_camo_ace",
-  "fow_h_us_m2_camo_heart",
-  "fow_h_us_m2_camo_diamond",
-  "fow_h_us_m2_camo_clover",
-  "fow_h_us_m2_camo_open",
-  "fow_h_us_m2_camo_open_ace",
-  "fow_h_us_m2_camo_open_heart",
-  "fow_h_us_m2_camo_open_diamond",
-  "fow_h_us_m2_camo_open_clover",
-  "fow_h_us_m2_fak",
-  "fow_h_us_m2_fak_net",
-  "fow_h_us_m2_fak_camo",
-  "fow_h_us_flight_helmet",
-  "fow_h_us_daisy_mae_01",
-  "fow_h_us_daisy_mae_02",
-  "fow_h_us_daisy_mae_03",
-  "fow_h_usmc_m1",
-  "fow_h_usmc_m1_camo_01",
-  "fow_h_usmc_m1_camo_02",
-  "fow_h_uk_mk2",
-  "fow_h_uk_mk2_net",
-  "fow_h_uk_mk2_net_camo",
-  "fow_h_uk_mk2_net_foliage",
-  "fow_h_uk_mk3",
-  "fow_h_uk_mk3_net_camo",
-  "fow_h_uk_mk2_para",
-  "fow_h_uk_mk2_para_camo",
-  "fow_h_uk_pol_mk2_para",
-  "fow_h_uk_pol_mk2_para_camo",
-  "fow_h_uk_mk2_para_foliage",
-  "fow_h_uk_woolen_hat",
-  "fow_h_uk_woolen_hat02",
-  "fow_h_uk_beret_para",
-  "fow_h_uk_beret_para_2",
-  "fow_h_uk_beret_royalmarines",
-  "fow_h_uk_beret_rtr",
-  "fow_h_uk_beret_rtr_headset",
-  "fow_h_uk_beret_rha",
-  "fow_h_uk_beret_rha_headset",
-  "fow_h_uk_beret_commando",
-  "fow_h_uk_beret_commando_kieffer",
-  "fow_h_uk_beret_commando_nor",
-  "fow_h_uk_beret_sas",
-  "fow_h_uk_beret_sas_2",
-  "fow_h_uk_beret_commando_pol",
-  "fow_h_uk_beret_para_pol",
-  "fow_h_uk_pib_beret",
-  "fow_h_uk_bp_beret",
-  "fow_h_uk_jungle_hat_01",
-  "fow_h_uk_jungle_hat_02",
-  "fow_h_uk_jungle_hat_03",
-
-  //IFA3
-  "H_LIB_GER_FSJ_M38_Helmet",
-  "H_LIB_GER_FSJ_M38_Helmet_os",
-  "H_LIB_GER_FSJ_M38_Helmet_grey",
-  "H_LIB_GER_FSJ_M38_Helmet_grey_os",
-  "H_LIB_GER_FSJ_M38_Helmet_DAK",
-  "H_LIB_GER_FSJ_M38_Helmet_Cover",
-  "H_LIB_GER_FSJ_M44_Helmet",
-  "H_LIB_GER_FSJ_M44_Helmet_os",
-  "H_LIB_GER_FSJ_M44_HelmetCamo1",
-  "H_LIB_GER_FSJ_M44_HelmetCamo2",
-  "H_LIB_GER_FSJ_M44_Helmet_Medic",
-  "H_LIB_GER_FSJ_M44_Helmet_Winter",
-  "H_LIB_GER_FSJ_M44_HelmetUtility",
-  "H_LIB_GER_TankOfficerCap",
-  "H_LIB_GER_TankOfficerCap2",
-  "H_LIB_GER_TankPrivateCap",
-  "H_LIB_GER_TankPrivateCap2",
-  "H_LIB_GER_SPGPrivateCap",
-  "H_LIB_GER_Helmet",
-  "H_LIB_GER_Helmet_net",
-  "H_LIB_GER_Helmet_ns",
-  "H_LIB_GER_Helmet_os",
-  "H_LIB_GER_Helmet_Medic",
-  "H_LIB_GER_Helmet_painted",
-  "H_LIB_GER_Helmet_net_painted",
-  "H_LIB_GER_Helmet_ns_painted",
-  "H_LIB_GER_Helmet_os_painted",
-  "H_LIB_GER_Helmet_Glasses",
-  "H_LIB_GER_HelmetUtility",
-  "H_LIB_GER_HelmetUtility_Grass",
-  "H_LIB_GER_HelmetUtility_Oak",
-  "H_LIB_GER_HelmetCamo",
-  "H_LIB_GER_HelmetCamo2",
-  "H_LIB_GER_HelmetCamo3",
-  "H_LIB_GER_HelmetCamo4",
-  "H_LIB_GER_OfficerCap",
-  "H_LIB_GER_Cap",
-  "H_LIB_GER_Fieldcap",
-  "H_LIB_GER_LW_PilotHelmet",
-  "H_LIB_GER_Ushanka",
-  "H_LIB_WP_Helmet",
-  "H_LIB_WP_Helmet_camo",
-  "H_LIB_WP_Helmet_med",
-  "H_LIB_WP_Cap",
-  "H_LIB_SOV_RA_PrivateCap",
-  "H_LIB_SOV_RA_OfficerCap",
-  "H_LIB_SOV_RA_Helmet",
-  "H_LIB_SOV_TankHelmet",
-  "H_LIB_SOV_PilotHelmet",
-  "H_LIB_SOV_Ushanka",
-  "H_LIB_SOV_Ushanka2",
-  "H_LIB_US_AB_Helmet",
-  "H_LIB_US_AB_Helmet_2",
-  "H_LIB_US_AB_Helmet_3",
-  "H_LIB_US_AB_Helmet_4",
-  "H_LIB_US_AB_Helmet_5",
-  "H_LIB_US_AB_Helmet_Jump_1",
-  "H_LIB_US_AB_Helmet_Jump_2",
-  "H_LIB_US_AB_Helmet_Clear_1",
-  "H_LIB_US_AB_Helmet_Clear_2",
-  "H_LIB_US_AB_Helmet_Clear_3",
-  "H_LIB_US_AB_Helmet_Medic_1",
-  "H_LIB_US_AB_Helmet_Plain_1",
-  "H_LIB_US_AB_Helmet_Plain_2",
-  "H_LIB_US_AB_Helmet_Plain_3",
-  "H_LIB_US_AB_Helmet_NCO_1",
-  "H_LIB_US_AB_Helmet_NCO_2",
-  "H_LIB_US_AB_Helmet_CO_1",
-  "H_LIB_US_AB_Helmet_CO_2",
-  "H_LIB_US_Helmet_Pilot",
-  "H_LIB_US_Helmet_Pilot_Glasses_Up",
-  "H_LIB_US_Helmet_Pilot_Glasses_Down",
-  "H_LIB_US_Helmet_Pilot_Respirator",
-  "H_LIB_US_Helmet_Pilot_Respirator_Glasses_Up",
-  "H_LIB_US_Helmet_Pilot_Respirator_Glasses_Down",
-  "H_LIB_US_Pilot_Cap",
-  "H_LIB_US_Pilot_Cap_Khaki",
-  "H_LIB_US_Helmet",
-  "H_LIB_US_Helmet_ns",
-  "H_LIB_US_Helmet_os",
-  "H_LIB_US_Helmet_Net",
-  "H_LIB_US_Helmet_Net_ns",
-  "H_LIB_US_Helmet_Net_os",
-  "H_LIB_US_Helmet_Med",
-  "H_LIB_US_Helmet_Med_ns",
-  "H_LIB_US_Helmet_Med_os",
-  "H_LIB_US_Helmet_Cap",
-  "H_LIB_US_Helmet_CO",
-  "H_LIB_US_Helmet_NCO",
-  "H_LIB_US_Helmet_First_lieutenant",
-  "H_LIB_US_Helmet_Second_lieutenant",
-  "H_LIB_US_Helmet_Tank",
-  "H_LIB_US_Rangers_Helmet",
-  "H_LIB_US_Rangers_Helmet_ns",
-  "H_LIB_US_Rangers_Helmet_os",
-  "H_LIB_US_Rangers_Helmet_Cap",
-  "H_LIB_US_Rangers_Helmet_First_lieutenant",
-  "H_LIB_US_Rangers_Helmet_Second_lieutenant",
-  "H_LIB_US_Rangers_Helmet_NCO",
-  "H_LIB_CIV_Worker_Cap_1",
-  "H_LIB_CIV_Worker_Cap_2",
-  "H_LIB_CIV_Worker_Cap_3",
-  "H_LIB_CIV_Worker_Cap_4",
-  "H_LIB_CIV_Villager_Cap_1",
-  "H_LIB_CIV_Villager_Cap_2",
-  "H_LIB_CIV_Villager_Cap_3",
-  "H_LIB_CIV_Villager_Cap_4",
-  "H_LIB_DAK_PithHelmet",
-  "H_LIB_DAK_Helmet",
-  "H_LIB_DAK_Helmet_2",
-  "H_LIB_DAK_Helmet_net",
-  "H_LIB_DAK_Helmet_net_2",
-  "H_LIB_DAK_Helmet_ns",
-  "H_LIB_DAK_Helmet_ns_2",
-  "H_LIB_DAK_Helmet_Glasses",
-  "H_LIB_DAK_OfficerCap",
-  "H_LIB_DAK_Cap",
-  "H_LIB_NKVD_PrivateCap",
-  "H_LIB_NKVD_OfficerCap",
-  "H_LIB_GER_Helmet_w",
-  "H_LIB_GER_Helmet_net_w",
-  "H_LIB_GER_Helmet_ns_w",
-  "H_LIB_GER_HelmetCamo_w",
-  "H_LIB_GER_HelmetCamob_w",
-  "H_LIB_GER_HelmetCamo2_w",
-  "H_LIB_GER_HelmetCamo2b_w",
-  "H_LIB_GER_HelmetCamo4_w",
-  "H_LIB_GER_Helmet_Glasses_w",
-  "H_LIB_GER_Cap_w",
-  "H_LIB_SOV_RA_Helmet_w",
-  "H_LIB_US_Helmet_w",
-  "H_LIB_US_Helmet_Cover_w",
-  "H_LIB_US_Helmet_Net_w",
-  "H_LIB_US_Helmet_Med_w",
-  "H_LIB_US_Helmet_Cap_w",
-  "H_LIB_US_Helmet_First_lieutenant_w",
-  "H_LIB_US_Helmet_Second_lieutenant_w",
-
-  //SAM3
-  "SAM_HelmetB_M90b",
-  "SAM_HelmetB_M90",
-  "SAM_HelmetB_M90K",
-  "SAM_HelmetIA_M90",
-  "SAM_HelmetIA_M90K"
-];
