--- addons/grad_slinghelmet/config.cpp
@@ -1,15 +1,0 @@
-class CfgPatches
-{
-	class GRAD_slingHelmet
-	{
-		units[] = {"GRAD_slingHelmet_WeaponHolder"};
-		weapons[] = {};
-		version = 0.1;
-		requiredVersion = 0.1;
-		requiredAddons[] = {"A3_Weapons_F","ace_common","ace_main","cba_settings"};
-		author = "DerZade [A]";
-	};
-};
-
-#include "CfgVehicles.hpp"
-#include "CfgFunctions.hpp"
