+++ addons/main/$PBOPREFIX$
@@ -1,0 +1,1 @@
+x\grad_slingHelmet\addons\main
