+++ addons/main/script_version.hpp
@@ -1,0 +1,4 @@
+#define MAJOR 1
+#define MINOR 2
+#define PATCHLVL 0
+#define BUILD 0
